+++ components/layout/mediaQuery/src/breakpoints.js
@@ -1,0 +1,8 @@
+export const BREAKPOINTS = {
+  xs: 480,
+  s: 600,
+  m: 840,
+  l: 960,
+  xl: 1280,
+  xxl: 1440
+}