+++ components/collapsible/basic/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="2.3.0"></a>
+# 2.3.0 (2018-08-03)
+
+
+### Features
+
+* added is clickable prop ([3ef14b9](https://github.com/SUI-Components/sui-components/commit/3ef14b9))
+
+
+
 <a name="2.2.0"></a>
 # 2.2.0 (2018-02-22)