+++ components/error/appBoundary/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="2.0.0"></a>
+# 2.0.0 (2018-08-02)
+
+
+### Bug Fixes
+
+* import correct styling from new component ([f870e72](https://github.com/SUI-Components/sui-components/commit/f870e72))
+
+
+### Features
+
+* use new React Molecule Notification component instead deprecated Alert Basi ([9ac8e6b](https://github.com/SUI-Components/sui-components/commit/9ac8e6b))
+
+
+### BREAKING CHANGES
+
+* Component not longer support icon property
+
+
+
 <a name="1.4.0"></a>
 # 1.4.0 (2018-01-22)
 
@@ -30,6 +50,16 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 * ensure the position of the error notification ([a84600c](https://github.com/SUI-Components/sui-components/commit/a84600c))
+
+
+
+<a name="1.1.0"></a>
+# 1.1.0 (2018-01-15)
+
+
+### Features
+
+* create new error appBoundary component ([98b0087](https://github.com/SUI-Components/sui-components/commit/98b0087))